Move provider selection back to domain
