--- src/app/[locale]/(common-layout)/_components/header/locale-selector/add-translate-dialog/can-translate.ts
@@ -1,6 +1,0 @@
-export function canTranslateWithoutGeminiApiKey(
-	hasGeminiApiKey: boolean,
-	selectedModel: string,
-) {
-	return hasGeminiApiKey || !selectedModel.startsWith("gemini-");
-}
