+++ src/app/[locale]/(common-layout)/_components/header/locale-selector/add-translate-dialog/client.tsx
@@ -23,8 +23,14 @@ import {
 	SelectValue,
 } from "@/components/ui/select";
 import { type TranslateActionState, translateAction } from "./action";
-import { canTranslateWithoutGeminiApiKey } from "./can-translate";
 import { DialogLocaleSelector } from "./dialog-locale-selector";
+
+export function canTranslateWithoutGeminiApiKey(
+	hasGeminiApiKey: boolean,
+	selectedModel: string,
+) {
+	return hasGeminiApiKey || !selectedModel.startsWith("gemini-");
+}
 
 type AddTranslateDialogProps = {
 	open: boolean;
